+++ ApiBaseReserva.Domain/Interfaces/Services/Common/IBaseService.cs
@@ -6,9 +6,9 @@ namespace ApiBaseReserva.Domain.Interfaces.Services.Common
     public interface IBaseService<TEntity> where TEntity : BaseEntity
     {
         TEntity Add(TEntity entity);
-        void Delete(int id);
+        void Delete(long id);
         IEnumerable<TEntity> GetAll();
-        TEntity GetById(int id);
+        TEntity GetById(long id);
         TEntity Update(TEntity entity);
     }
 }
